Implementation Of Patch Mapping

+++ src/main/java/com/springboot/web/services/EmployeeService.java
@@ -4,9 +4,12 @@ import com.springboot.web.dto.EmployeeDTO;
 import com.springboot.web.entities.EmployeeEntity;
 import com.springboot.web.repositories.EmployeeRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.util.ReflectionUtils;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.Field;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -52,10 +55,28 @@ private final ModelMapper modelMapper;
         EmployeeEntity savedEmployeeEntity = employeeRepository.save(employeeEntity);
         return modelMapper.map(savedEmployeeEntity, EmployeeDTO.class);
     }
-
-    public void deleteById(Long employeeId) {
-   employeeRepository.deleteById(employeeId);
-   System.out.println("Item is Succesfully Deleted");
+public boolean isEmployeeexistById(Long employeeId)
+{
+    return employeeRepository.existsById(employeeId);
+}
+    public boolean deleteById(Long employeeId) {
+   boolean exist=isEmployeeexistById(employeeId);
+   if(!exist) return false;
+        employeeRepository.deleteById(employeeId);
+   return true;
     }
 
+    public EmployeeDTO updatePartialEmployeeById(Map<String, Object> updates, Long employeeId) {
+        boolean exist=isEmployeeexistById(employeeId);
+        if(!exist) return null;
+        EmployeeEntity employeeEntity=employeeRepository.findById(employeeId).get();
+        updates.forEach((field,value) ->
+                {
+                    Field field1=ReflectionUtils.findRequiredField(EmployeeEntity.class,field);
+                     field1.setAccessible(true);
+                   ReflectionUtils.setField(field1,employeeEntity,value);
+                }
+                );
+return  modelMapper.map(employeeRepository.save(employeeEntity),EmployeeDTO.class);
+    }
 }
